Rename AuthGuard.ts to auth-guard.ts for better consistency
